refactor: AwsS3 -> image 이름 변경
