+++ src/main/java/bokjak/bokjakserver/domain/image/exception/ImageException.java
@@ -1,4 +1,4 @@
-package bokjak.bokjakserver.util.s3.exception;
+package bokjak.bokjakserver.domain.image.exception;
 
 import bokjak.bokjakserver.common.exception.StatusCode;
 import lombok.Getter;
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @Getter
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class AwsS3Exception extends RuntimeException{
+public class ImageException extends RuntimeException{
     StatusCode awsS3ErrorCode;
 
-    public AwsS3Exception(StatusCode awsS3ErrorCode) {
+    public ImageException(StatusCode awsS3ErrorCode) {
         super(awsS3ErrorCode.getMessage());
         this.awsS3ErrorCode = awsS3ErrorCode;
     }
