+++ src/main/java/bokjak/bokjakserver/domain/spot/service/SpotService.java
@@ -17,8 +17,8 @@ import bokjak.bokjakserver.domain.spot.model.SpotImage;
 import bokjak.bokjakserver.domain.spot.repository.SpotRepository;
 import bokjak.bokjakserver.domain.user.model.User;
 import bokjak.bokjakserver.domain.user.service.UserService;
-import bokjak.bokjakserver.util.s3.S3SaveDir;
-import bokjak.bokjakserver.util.s3.service.AwsS3Service;
+import bokjak.bokjakserver.domain.image.S3SaveDir;
+import bokjak.bokjakserver.domain.image.service.AwsS3Service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
